+++ src/Declination.cpp
@@ -1,4 +1,5 @@
-#include "../Configuration.hpp"  // For NORTHERN_HEMISPHERE only
+#include "./inc/Globals.hpp"
+#include "../Configuration.hpp"
 #include "Utility.hpp"
 #include "Declination.hpp"
 
@@ -11,13 +12,13 @@
 // Parses the RA or DEC from a string that has an optional sign, a two digit degree, a seperator, a two digit minute, a seperator and a two digit second.
 // For example:   -45*32:11 or 23:44:22
 
-// In the northern hemisphere, 0 is north pole, 180 and -180 is south pole
+// In the NORTHERN hemisphere, 0 is north pole, 180 and -180 is south pole
 //------------------ S---------------------------------------- N ---------------------------------- S
 // Celestial       -90    -60    -30      0      30    60     90    60    30    0    -30   -60    -90
 //                                    Celestial = 90 - abs(Declination)
 // Declination    -180    -150   -120    -90    -60   -30     0     30    60    90   120   150    180
 
-// In the southern hemisphere, 0 is south pole, 180 and -180 is north pole
+// In the SOUTHERN hemisphere, 0 is south pole, 180 and -180 is north pole
 //------------------ N---------------------------------------- S ---------------------------------- N
 // Celestial        90     60     30      0     -30   -60     -90    -60    -30    0    30   60    90
 //                                    Celestial = -90 + abs(Declination)
@@ -87,7 +88,9 @@ const char *Declination::ToString() const
 
     *p++ = ' ';
     *p++ = '(';
-    strcpy(p, String(NORTHERN_HEMISPHERE ? 90 - fabsf(getTotalHours()) : -90 + fabsf(getTotalHours()), 4).c_str());
+    strcpy(p, String(inNorthernHemisphere ? 90 - fabsf(getTotalHours()) : -90 + fabsf(getTotalHours()), 4).c_str());
+    strcat(p, ", ");
+    strcat(p, String(getTotalHours(), 4).c_str());
     strcat(p, ")");
 
     return achBufDeg;
@@ -96,15 +99,16 @@ const char *Declination::ToString() const
 Declination Declination::ParseFromMeade(String const &s)
 {
     Declination result;
-    LOG(DEBUG_GENERAL, "[DECLINATION]: Declination.Parse(%s)", s.c_str());
+    LOG(DEBUG_MEADE, "[DECLINATION]: Declination.Parse(%s) for %s Hemi", s.c_str(), inNorthernHemisphere ? "N" : "S");
 
     // Use the DayTime code to parse it...
     DayTime dt = DayTime::ParseFromMeade(s);
+    LOG(DEBUG_MEADE, "[DECLINATION]: Declination DayTime is %l secs", dt.getTotalSeconds());
 
     // ...and then correct for hemisphere
-    result.totalSeconds = NORTHERN_HEMISPHERE ? (arcSecondsPerHemisphere / 2) - dt.getTotalSeconds()
-                                              : -(arcSecondsPerHemisphere / 2) + dt.getTotalSeconds();
-    LOG(DEBUG_GENERAL, "[DECLINATION]: Declination.Parse(%s) -> %s (%l)", s.c_str(), result.ToString(), result.totalSeconds);
+    result.totalSeconds = inNorthernHemisphere ? (arcSecondsPerHemisphere / 2) - labs(dt.getTotalSeconds())
+                                               : -(arcSecondsPerHemisphere / 2) + labs(dt.getTotalSeconds());
+    LOG(DEBUG_MEADE, "[DECLINATION]: Adjust for hemisphere. %s -> %s (%l secs)", s.c_str(), result.ToString(), result.totalSeconds);
     return result;
 }
 
@@ -112,16 +116,16 @@ Declination Declination::FromSeconds(long seconds)
 {
     const auto secondsFloat                 = static_cast<float>(seconds);
     const auto arcSecondsPerHemisphereFloat = static_cast<float>(arcSecondsPerHemisphere);
-#if NORTHERN_HEMISPHERE == 1
-    return Declination(((arcSecondsPerHemisphereFloat / 2.0f) - secondsFloat) / 3600.0f);
-#else
-    return Declination(((-arcSecondsPerHemisphereFloat / 2.0f) + secondsFloat) / 3600.0f);
-#endif
+    if (inNorthernHemisphere)
+    {
+        return Declination(((arcSecondsPerHemisphereFloat / 2.0f) - secondsFloat) / 3600.0f);
+    }
+    return Declination(((arcSecondsPerHemisphereFloat / 2.0f) + secondsFloat) / 3600.0f);
 }
 
 const char *Declination::formatString(char *targetBuffer, const char *format, long *) const
 {
     long secs
-        = NORTHERN_HEMISPHERE ? (arcSecondsPerHemisphere / 2) - labs(totalSeconds) : -(arcSecondsPerHemisphere / 2) + labs(totalSeconds);
+        = inNorthernHemisphere ? (arcSecondsPerHemisphere / 2) - labs(totalSeconds) : -(arcSecondsPerHemisphere / 2) + labs(totalSeconds);
     return DayTime::formatString(targetBuffer, format, &secs);
 }